finished persisting cache
